AST+Indexer+Span+Syntax+Type: Implement Serde traits for various structures
