+++ crates/pxp-ast/src/enums.rs
@@ -7,6 +7,7 @@ use crate::identifiers::SimpleIdentifier;
 
 use crate::Expression;
 use pxp_span::Span;
+use pxp_syntax::backed_enum_type::BackedEnumType;
 
 #[derive(Debug, PartialEq, Eq, Clone)]
 
@@ -40,39 +41,6 @@ pub struct UnitEnumStatement {
     pub name: SimpleIdentifier,            // `Foo`
     pub implements: Vec<SimpleIdentifier>, // `implements Bar`
     pub body: UnitEnumBody,                // `{ ... }`
-}
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-
-pub enum BackedEnumType {
-    String(Span, Span), // `:` + `string`
-    Int(Span, Span),    // `:` + `int`
-    Invalid(Span),
-}
-
-impl BackedEnumType {
-    pub fn is_valid(&self) -> bool {
-        match self {
-            Self::String(..) | Self::Int(..) => true,
-            Self::Invalid(..) => false,
-        }
-    }
-}
-
-impl Display for BackedEnumType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            BackedEnumType::String(..) => write!(f, "string"),
-            BackedEnumType::Int(..) => write!(f, "int"),
-            BackedEnumType::Invalid(..) => write!(f, "invalid"),
-        }
-    }
-}
-
-impl Default for BackedEnumType {
-    fn default() -> Self {
-        Self::Invalid(Span::default())
-    }
 }
 
 #[derive(Debug, PartialEq, Eq, Clone)]
