+++ crates/pxp-ast/src/modifiers.rs
@@ -1,25 +1,7 @@
 use std::fmt::{Display, Formatter};
 
 use pxp_span::Span;
-
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Default)]
-
-pub enum Visibility {
-    #[default]
-    Public,
-    Protected,
-    Private,
-}
-
-impl Display for Visibility {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Visibility::Public => write!(f, "public"),
-            Visibility::Protected => write!(f, "protected"),
-            Visibility::Private => write!(f, "private"),
-        }
-    }
-}
+use pxp_syntax::visibility::Visibility;
 
 #[derive(Debug, PartialEq, Eq, Clone)]
