+++ crates/pxp-parser/src/internal/enums.rs
@@ -7,7 +7,6 @@ use pxp_ast::enums::BackedEnumBody;
 use pxp_ast::enums::BackedEnumCase;
 use pxp_ast::enums::BackedEnumMember;
 use pxp_ast::enums::BackedEnumStatement;
-use pxp_ast::enums::BackedEnumType;
 use pxp_ast::enums::UnitEnumBody;
 use pxp_ast::enums::UnitEnumCase;
 use pxp_ast::enums::UnitEnumMember;
@@ -16,6 +15,7 @@ use pxp_ast::StatementKind;
 use pxp_diagnostics::DiagnosticKind;
 use pxp_diagnostics::Severity;
 use pxp_span::Span;
+use pxp_syntax::backed_enum_type::BackedEnumType;
 use pxp_token::TokenKind;
 
 use super::classes::member;
