+++ crates/pxp-span/Cargo.toml
@@ -7,3 +7,4 @@ license-file.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+serde = { version = "1.0.193", features = ["derive"] }
