+++ crates/pxp-span/src/lib.rs
@@ -1,4 +1,6 @@
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Default)]
+use serde::{Serialize, Deserialize};
+
+#[derive(Debug, PartialEq, Eq, Clone, Copy, Default, Serialize, Deserialize)]
 pub struct Span {
     pub start: Position,
     pub end: Position,
@@ -18,7 +20,7 @@ impl Span {
     }
 }
 
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Default)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy, Default, Serialize, Deserialize)]
 pub struct Position {
     pub offset: usize,
     pub line: usize,
