+++ crates/pxp-syntax/Cargo.toml
@@ -9,3 +9,4 @@ license-file.workspace = true
 [dependencies]
 pxp-span = { path = "../pxp-span" }
 pxp-symbol = { path = "../pxp-symbol" }
+serde = { version = "1.0.193", features = ["derive"] }
