+++ crates/pxp-syntax/src/lib.rs
@@ -1,1 +1,3 @@
 pub mod comments;
+pub mod visibility;
+pub mod backed_enum_type;
